requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.6
et-xmlfile==1.1.0
idna==3.2
numpy==1.21.2
openpyxl==3.0.9
pandas==1.3.3
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
urllib3==1.26.7
